requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

pdfplumber==0.10.4
python-docx==1.1.0

openai==1.10.0

python-multipart==0.0.6

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

black==24.1.1
isort==5.13.2
ruff==0.1.14
pylint==3.0.3
mypy==1.8.0
